requirements.txt
fastapi
uvicorn
sqlalchemy
async-exit-stack
async-generator
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
numpy
pandas
seaborn
matplotlib
plotly
emoji
scikit-learn
deta
pandas
joblib
mypy
sqlalchemy
mysql-connector-python
python-dotenv
gunicorn